Use Modulino/RouterBridge; update collector and DB
Replace MPU-6050 sketch with a Modulino Movement (LSM6DSOX) sketch that uses the Arduino Router Bridge Monitor (moved under arduino/schlafschaf/). Add CLAUDE.md documentation describing project architecture and usage. Update python/collector.py to support an on-device RouterBridgePort (defaulting to /var/run/arduino-router.sock / 127.0.0.1:7500), parse IMU values as floats, compute magnitude as float, and select between TCP/serial ports (with improved reconnect logic). Update python/database.py to store accel/gyro columns as REAL and adjust insert_raw types accordingly.

+++ python/database.py
@@ -43,12 +43,12 @@ class Database:
                 session_id TEXT NOT NULL,
                 timestamp REAL NOT NULL,
                 sound INTEGER,
-                ax INTEGER,
-                ay INTEGER,
-                az INTEGER,
-                gx INTEGER,
-                gy INTEGER,
-                gz INTEGER,
+                ax REAL,
+                ay REAL,
+                az REAL,
+                gx REAL,
+                gy REAL,
+                gz REAL,
                 FOREIGN KEY (session_id) REFERENCES sessions(id)
             );
 
@@ -143,8 +143,8 @@ class Database:
     # ── Rohdaten ─────────────────────────────────────────────────────────────
 
     def insert_raw(self, session_id: str, timestamp: float,
-                   ax: int, ay: int, az: int,
-                   gx: int, gy: int, gz: int,
+                   ax: float, ay: float, az: float,
+                   gx: float, gy: float, gz: float,
                    sound: int | None = None):
         self.conn.execute(
             """INSERT INTO raw_data
